checking if form submitted and not deleting

+++ quast_app/views.py
@@ -820,16 +820,19 @@ def delete_session(request):
         logger.error('quast_app.views.delete_session: No quast session with report_id=%s' % report_id)
         return HttpResponseBadRequest('wrong reportId: no such quast-session')
 
-    if quast_session.contigs_files:
-        fpaths = [os.path.join(quast_session.get_contigs_dirpath(), c_f.fname) for c_f in quast_session.contigs_files.all()]
-        for fpath in fpaths:
-            if os.path.exists(fpath):
-                try:
-                    os.remove(fpath)
-                    logger.info('quast_app.views.delete_session: Deleted contigs_file %s' % fpath)
-                except Exception, e:
-                    logger.warn('quast_app.views.delete_session: Error deleting contigs file %s: %s' % (fpath, e.message))
+    if quast_session.submited:
+        if quast_session.contigs_files:
+            fpaths = [os.path.join(quast_session.get_contigs_dirpath(), c_f.fname) for c_f in quast_session.contigs_files.all()]
+            for fpath in fpaths:
+                if os.path.exists(fpath):
+                    try:
+                        os.remove(fpath)
+                        logger.info('quast_app.views.delete_session: Deleted contigs_file %s' % fpath)
+                    except Exception, e:
+                        logger.warn('quast_app.views.delete_session: Error deleting contigs file %s: %s' % (fpath, e.message))
 
-    logger.info('quast_app.views.delete_session: Deleting quast_session with id=%s' % report_id)
-    quast_session.delete()
+        logger.info('quast_app.views.delete_session: Deleting quast_session with id=%s' % report_id)
+        quast_session.delete()
+    else:
+        logger.info('quast_app.views.delete_session: Quast session with id=%s wasn\'t submitted, not deleting' % report_id)
     return HttpResponse()
